submit_color now doesn't fail
